test: move `EARL` and `RDFT` namespaces to separate files
This patch moves the `RDFT` and `EARL` namespaces into their own files.

This is to make the test code a bit cleaner to work with. Doing this in
preperation for rewrite of the RDF/XML test suite.
