+++ test/test_w3c_spec/test_nt_w3c.py
@@ -3,7 +3,8 @@ test suite."""
 import os
 from test import TEST_DIR
 from test.data import TEST_DATA_DIR
-from test.utils.manifest import RDFT, RDFTest, read_manifest
+from test.utils.manifest import RDFTest, read_manifest
+from test.utils.namespace import RDFT
 from typing import Callable, Dict
 
 import pytest
