+++ test/test_w3c_spec/test_turtle_w3c.py
@@ -5,7 +5,8 @@ import os
 from pathlib import Path
 from test.data import TEST_DATA_DIR
 from test.utils import file_uri_to_path
-from test.utils.manifest import RDFT, RDFTest, read_manifest
+from test.utils.manifest import RDFTest, read_manifest
+from test.utils.namespace import RDFT
 from typing import Callable, Dict, Set
 
 import pytest
