+++ test/utils/namespace.py
@@ -1,0 +1,7 @@
+from ._EARL import EARL
+from ._RDFT import RDFT
+
+__all__ = [
+    "EARL",
+    "RDFT",
+]
